Move run code to separate file
