+++ manual/test.py
@@ -1,6 +1,4 @@
 # Generated by Selenium IDE
-import concurrent.futures
-from sys import argv
 import pytest
 import time
 import json
@@ -364,40 +362,3 @@ class TestDefaultSuite():
     self.vars["country"] = "UnitedKingdom"
     # 4 | close |  | 
     self.driver.close()
-  
-def run_one_country(country):
-    suite = TestDefaultSuite()
-    suite.setup_method()
-    getattr(suite, 'test_'  + country)()
-    return suite.vars
-
-if __name__ == '__main__':
-    country_list = []
-    countries = []
-    for m in dir(TestDefaultSuite):
-        if m.startswith('test_'):
-            countries.append(m[5:])
-    if argv[1:]:
-        # check specified country names are valid
-        for a in argv[1:]:
-            if a not in countries:
-                print('USAGE:')
-                print('\t%s <country1> <country2> ...' % argv[0])
-                print('\t- to test country1 country2 ....')
-                print('or')
-                print('\t%s' % argv[0])
-                print('\t- to test all countries')
-                print('Valid country names:')
-                print('\n'.join(countries))
-                assert a in countries, a + ' not in ' + str(countries)
-        countries = set(argv[1:])
-        for country in countries:
-            country_list.append(run_one_country(country))
-    else:
-        # run all countries in parallel
-        with concurrent.futures.ThreadPoolExecutor() as executor:
-            country_list = executor.map(run_one_country, countries)
-    with open("list_data.txt", 'w', encoding="utf-8") as f:
-                                for item in country_list:
-                                    f.write("%s\n" % item)
-                                f.close()
